split off mail class
